Separate BuilderExtensions
